requirements.txt
# Core dependencies for LLM extraction pipeline
orkg>=1.0.0
openai>=1.0.0  # For KISSKI API (OpenAI-compatible)
arxiv>=2.0.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
sentence-transformers>=2.2.0  # For semantic similarity in evaluation
bert-score>=0.3.13  # For BERTScore semantic evaluation (thesis metrics)